requirements.txt
scipy
numpy
pyserial
seaborn
